Added a service helper. Updated return types from string list to string array
